% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_dataset.R
\name{access_dataset}
\alias{access_dataset}
\title{Get dataset overview function}
\usage{
access_dataset()
}
\description{
This function gets the most recent metadata from the file servers and stores it as a variable
}
\examples{
access_dataset()

}
\keyword{data}
