% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_corpus_search.R
\name{do_corpus_search}
\alias{do_corpus_search}
\title{Do corpus search}
\usage{
do_corpus_search(
  searchterm = "oskar kallas",
  searchfile = "results1.txt",
  filelist = filelist
)
}
\arguments{
\item{searchterm}{Write the search term here. By default it is read as regular expressions.}

\item{searchfile}{Write the filename to store the results in.}

\item{filelist}{It needs the filelist for raw texts as input too.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
do_corpus_search("oskar kallas","results1.txt")

}
\keyword{data}
