% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_subset_search.R
\name{do_subset_search}
\alias{do_subset_search}
\title{Do corpus search}
\usage{
do_subset_search(
  searchterm = "oskar kallas",
  searchfile = "oskarkallas.txt",
  subset
)
}
\arguments{
\item{searchterm}{Write the search term here. By default it is read as regular expressions.}

\item{searchfile}{Write the filename to store the results in.}

\item{subset}{Needs subset to be specified}
}
\description{
This function performs a search within the subset of the corpus files
}
\examples{
do_subset_search("oskar kallas","results1.txt",subset)

}
\keyword{data}
