% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concordances.R
\name{get_concordances}
\alias{get_concordances}
\title{Get concordances}
\usage{
get_concordances(
  searchterm,
  texts,
  before = 30,
  after = 30,
  txtcolumn = "txt",
  idcolumn = "id"
)
}
\arguments{
\item{searchterm}{Write the search term here. By default it is read as regular expressions.}

\item{searchfile}{Write the filename to store the results in.}

\item{subset}{Needs subset to be specified}
}
\description{
This function gets he concordances.
}
\examples{
get_concordances("searchterm",texts,30,30,"txt","id")

}
\keyword{data}
