% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus_filelists.R
\name{get_corpus_filelists}
\alias{get_corpus_filelists}
\title{Get corpus filelists}
\usage{
get_corpus_filelists(subset)
}
\arguments{
\item{subset}{Needs a subset file as input. Make this by selecting the required rows from the whole metadata file.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get_corpus_filelists(subset)

}
\keyword{data}
