% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus_meta.R
\name{get_corpus_meta}
\alias{get_corpus_meta}
\title{Get corpus metadata}
\usage{
get_corpus_meta(metafilelist)
}
\arguments{
\item{metafilelist}{Needs to get the metafilelist from the subset.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
get_corpus_meta()

}
\keyword{data}
