% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_digar_overview.R
\name{get_digar_overview}
\alias{get_digar_overview}
\title{Get dataset overview function}
\usage{
get_digar_overview()
}
\description{
This function gets the most recent metadata from the file servers and stores it as a variable
}
\examples{
all_issues <- get_digar_overview()

}
\keyword{data}
