% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matching_files.R
\name{get_matching_files}
\alias{get_matching_files}
\title{Do corpus search}
\usage{
get_matching_files(
  fileids_w_meta,
  outputfile = "file2.txt",
  source = "sections",
  extracttype = "text"
)
}
\arguments{
\item{fileids_w_meta}{the table to use for fileids.Must have zippath_sections, zippath_pages, keyid, year, month.}

\item{outputfile}{Write the filename to store the results in.}

\item{source}{Either "sections" or "pages" based on what kind of input is expected.}

\item{extracttype}{Either "text" or "lemmas" based on what kind of texts you want to extract.}
}
\description{
This function gets either text or lemma data based on ids
}
\examples{
get_matching_files(fileds_w_meta, "results2.txt","sections","text")

}
\keyword{data}
