% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subset_meta.R
\name{get_subset_meta}
\alias{get_subset_meta}
\title{Get corpus metadata}
\usage{
get_subset_meta(subset)
}
\arguments{
\item{subset}{Needs to get the subset from the subset.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
subset_meta <- get_subset_meta(subset)

}
\keyword{data}
